+++ backend/Library_API.Application/UseCases/BookUseCases/DeleteBookUseCase.cs
@@ -1,7 +1,6 @@
 ﻿using Library_API.Application.UseCases.BookUseCases.BooksUseCasesInterfaces;
 using Library_API.Core.Abstractions;
 using Library_API.Core.Models;
-using Library_API.DataAccess.Repositories;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,9 +18,14 @@ namespace Library_API.Application.UseCases.BookUseCases
             _booksRepository = booksRepository;
         }
 
-        public async Task<Guid> ExecuteAsync(Guid id)
+        public async Task ExecuteAsync(Guid id)
         {
-            return await _booksRepository.Delete(id);
+            bool isExist = await _booksRepository.IsExist(id);
+            if (!isExist)
+            {
+                throw new Exception("Book does not exist");
+            }
+            await _booksRepository.Delete(id);
         }
     }
 }
